django-crud-api/pyproject.toml
[tool.poetry]
name = "django-crud-api"
version = "0.1.0"
description = ""
authors = ["João Paulo Carvalho <jjpaulo2@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
drf-yasg = "^1.21.4"

[tool.poetry.dev-dependencies]
pylint = "^2.15.5"
jedi = "^0.18.1"
bandit = "^1.7.4"
mypy = "^0.982"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
disable = [
    "missing-docstring"
]